Add files via upload
Two Important changes to note in this commit. In the large code block it is set up to run the first 10 streets in my list of streets. You will need to change the first line in the for loop to reflect the whole data set. I have not explored much of the dataset beyond the first ten streets. Additionally, in just that limited scope I have gotten duplicate data in an unusual pattern that I think may be on the network side. Sometime I get 8 duplicate parcel ids other times I get 10 or more. Have fun and I'll try to start pulling more data!